flask/requirements.txt
flask 
uwsgi
flask-sqlalchemy
flask-migrate
flask-script
flask-marshmallow
flask-bcrypt 
pyjwt
psycopg2-binary
flask-restful 
flask-jwt-extended
marshmallow-sqlalchemy
flask_cors
python-dotenv
